separate activation to its own file
